dialog box for username reset working
