+++ src/login/login.component.ts
@@ -14,13 +14,12 @@ import {
 } from "@angular/forms";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatInputModule } from "@angular/material/input";
-import {
-  MatDialog,
-  MatDialogModule,
-  MatDialogRef
-} from "@angular/material/dialog";
+import { MatDialog, MatDialogModule } from "@angular/material/dialog";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { MatButtonModule } from "@angular/material/button";
+
+import { PassReset } from "src/login/password-reset";
+import { UserReset } from "src/login/username-reset";
 
 @Component({
   selector: "login",
@@ -29,6 +28,7 @@ import { MatButtonModule } from "@angular/material/button";
   encapsulation: ViewEncapsulation.None
 })
 @NgModule({
+  declarations: [PassReset, UserReset],
   imports: [
     BrowserAnimationsModule,
     ReactiveFormsModule,
@@ -38,6 +38,7 @@ import { MatButtonModule } from "@angular/material/button";
     MatFormFieldModule,
     MatInputModule
   ],
+  entryComponents: [PassReset, UserReset],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class GameLogin implements OnInit {
@@ -46,10 +47,15 @@ export class GameLogin implements OnInit {
 
   constructor(private formBuilder: FormBuilder, private dialog: MatDialog) {}
 
-  openDialog() {
+  openPasswordDialog() {
     const dialogRef = this.dialog.open(PassReset);
     dialogRef.afterClosed().subscribe(() => console.log("Dialog box closed"));
   }
+  openUsernameDialog() {
+    const dialogRef = this.dialog.open(UserReset);
+    dialogRef.afterClosed().subscribe(() => console.log("Dialog box closed"));
+  }
+
   ValidateUser(control: AbstractControl): { [key: string]: any } | null {
     if (control.value !== "goship97") {
       return { userValid: true };
@@ -77,50 +83,5 @@ export class GameLogin implements OnInit {
       return window.alert("Needs valid input");
     }
     console.log(this.gameForm.value);
-  }
-}
-
-@Component({
-  selector: "password-reset",
-  templateUrl: "password-reset.html",
-  styleUrls: ["password-reset.css"]
-})
-export class PassReset implements OnInit {
-  passwordTitle = "Password Reset";
-  resetInfo: string = "Enter username below: ";
-  sendButtonText: string = "Send";
-  closeButtonText: string = "Close";
-  successMessage: string = "";
-
-  passResetForm: FormGroup;
-  passResetSubmit = false;
-
-  constructor(
-    private passFormBuilder: FormBuilder,
-    private dialogRef: MatDialogRef<PassReset>
-  ) {}
-
-  ResetPassValidate(control: AbstractControl): { [key: string]: any } | null {
-    if (control.value !== "goship97") {
-      return { resetPassValid: true };
-    }
-    return null;
-  }
-
-  ngOnInit() {
-    this.passResetForm = this.passFormBuilder.group({
-      passResetText: ["", [this.ResetPassValidate]]
-    });
-  }
-  sendInfo() {
-    this.passResetSubmit = true;
-
-    if (this.passResetForm.invalid) {
-      return (this.successMessage = "Email not sent. Try a valid username.");
-    }
-    this.successMessage = "Email sent!";
-  }
-  onNoClick(): void {
-    this.dialogRef.close(true);
   }
 }
